requirements.txt
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
tensorflow>=2.15.0
keras>=3.0.0
opencv-python-headless>=4.10.0
Pillow>=10.0.0
Flask>=3.0.0
Werkzeug>=3.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
h5py>=3.10.0
gunicorn>=22.0.0